+++ changelog.md
@@ -5,6 +5,7 @@
 - added new operators: and, or, nand, nor, =:, sum, prod 
 - removed operator: getattr
 - rewritten operator: xor, xnor (now using \_\_bool__ instead of \_\_eq__)
+- classes do no longer have their own namespace
 
 # Version 0.2.8
 - added FOREACH loops with automatically zip() when multiple iterable are given